4.15-4.23

+++ osa4/blogilista/controllers/blogs.js
@@ -1,8 +1,10 @@
 const blogsRouter = require("express").Router();
+const { userExtractor } = require("../utils/middleware.js");
 const Blog = require("../models/blog");
+const User = require("../models/user");
 
 blogsRouter.get("/", async (request, response) => {
-  const blogs = await Blog.find({});
+  const blogs = await Blog.find({}).populate("user");
   response.json(blogs);
 });
 
@@ -15,23 +17,39 @@ blogsRouter.get("/:id", async (request, response) => {
   }
 });
 
-blogsRouter.post("/", async (request, response) => {
-  const blog = new Blog(request.body);
+blogsRouter.post("/", userExtractor, async (request, response) => {
+  const user = await User.findById(request.user);
+  const blog = new Blog({ ...request.body, user: user._id });
 
   if (!blog.title || !blog.url) {
     response.status(400).end();
+  } else if (request.user !== user._id.toString()) {
+    response.status(403).json({ error: "unauthorized access" });
   } else {
     const result = await blog.save();
+
+    user.blogs = user.blogs.concat(result._id);
+    user.save();
+
     response.status(201).json(result);
   }
 });
 
-blogsRouter.delete("/:id", async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id);
-  response.status(204).end();
+blogsRouter.delete("/:id", userExtractor, async (request, response) => {
+  const blog = await Blog.findById(request.params.id);
+
+  if (blog.user.toString() === request.user) {
+    await Blog.findByIdAndDelete(request.params.id);
+    await User.findByIdAndUpdate(request.user, {
+      $pull: { blogs: request.params.id },
+    });
+    response.status(204).end();
+  } else {
+    response.status(403).json({ error: "unauthorized access" });
+  }
 });
 
-blogsRouter.put("/:id", async (request, response) => {
+blogsRouter.put("/:id", userExtractor, async (request, response) => {
   const updated = await Blog.findByIdAndUpdate(
     request.params.id,
     request.body,
